Move methods related to params to parameters.rb
