+++ app/controllers/api/base_controller/parameters.rb
@@ -1,6 +1,28 @@
 module Api
   class BaseController
     module Parameters
+      module ResultsController
+        def sort_order
+          params['sort_order'] == 'desc' ? :descending : :ascending
+        end
+
+        def param_result_set?
+          params.key?(:hash_attribute) && params[:hash_attribute] == "result_set"
+        end
+
+        def report_options
+          params.merge(:sort_by => params['sort_by'], :sort_order => sort_order).merge(filter_options)
+        end
+
+        def filter_options
+          filtering_enabled? ? {:filter_string => params[:filter_string], :filter_column => params[:filter_column]} : {}
+        end
+
+        def filtering_enabled?
+          params.key?(:filter_column) && params.key?(:filter_string) && params[:filter_string]
+        end
+      end
+
       def hash_fetch(hash, element, default = {})
         hash[element] || default
       end
